% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bac.R
\docType{data}
\name{bac}
\alias{bac}
\title{SNVs after BCF tool filtered MpileUp VCF data.}
\format{A VCF object imported with readVcf function from VariantAnnotation package}
\usage{
bac
}
\description{
A dataset containing the SNV positions and their alternative nucleotide frequencies
of three Illumina sequenced baculoviruses.
}
\keyword{datasets}
