% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacsnp.R
\docType{package}
\name{bacsnp-package}
\alias{bacsnp}
\alias{bacsnp-package}
\title{bacsnp: A package for SNP frequencies of baculoviruses}
\description{
The bacsnp package provides four helpful functions for SNP analysis from VCF files:
vcf_transformation, filter_snp_df, snp_specificity, and bacsnp.plot.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/wennj/bacsnp}
  \item Report bugs at \url{http://github.com/wennj/bacsnp/issues}
}

}
\author{
\strong{Maintainer}: Jörg Wennmann \email{bacsnp@gmail.com}

Authors:
\itemize{
  \item Jiangbin Fan
}

Other contributors:
\itemize{
  \item Sergei Senger [contributor]
}

}
\keyword{internal}
