% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_plot.R
\name{bacsnp.plot}
\alias{bacsnp.plot}
\title{SNP frequency plot}
\usage{
bacsnp.plot(x, col = "SPEC", genome.length = NULL,
  mark.repeats = NULL, mark.lowGQ = NULL, mark.lowQUAL = NULL)
}
\arguments{
\item{x}{A VCF dataframe that contains the SNP data from a single isolate only.
vcf_transformation.}

\item{col}{Column of the dataframe which is used to colorize each SNP position.}

\item{genome.length}{The length of the reference genome. If set to default the last SNP position is used.}

\item{mark.repeats}{Dataframe that contains the repeat regions of the genome.}

\item{mark.lowGQ}{Numeric value >= 0}

\item{mark.lowQUAL}{Numeric value >= 0}
}
\value{
A list of two elements. First element is the input dataframe with SNP
  specificities added. The second element is a counting talbe with all SNP
  specificity combinations.
}
\description{
Function to draw a SNP frequency plot. Requires ggplot2.
}
\examples{
iso1 <- bac[bac$ISO == "iIsolate1",]
bacsnp.plot(iso1, col = "SPEC")
bacsnp.plot(x, col = "SPEC", genome.length = 125000)
}
