% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_specificity.R
\name{bacsnp.specificity}
\alias{bacsnp.specificity}
\title{SNP specificity determination}
\usage{
bacsnp.specificity(vcfdf, isolates, which.rel = "REL.ALT1")
}
\arguments{
\item{vcfdf}{A dataframe with SNP frequencies. Data was created with vcf_transformation.}

\item{isolates}{A character vector for which the SNP specificity determination is performed.}

\item{which.rel}{Specifies the column of relativ SNP frequencies for which the specificity determination should be performed on.}
}
\value{
A list of two elements. First element is the input dataframe with SNP specificities added. The second element is a counting talbe with all SNP specificity combinations.
}
\description{
SNP specificity determination
}
\examples{
data(bac.f)
bacsnp.specificity(bac.f, c("iIso1", "iIso2"), which.rel = "REL.ALT1")
}
