% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_transformation.R
\name{bacsnp.transformation}
\alias{bacsnp.transformation}
\title{Transforming a VCF file into a dataframe.}
\usage{
bacsnp.transformation(vcf)
}
\arguments{
\item{vcf}{A VCF file loaded with readVCF of the VariantAnnotation package.}
}
\value{
A dataframe.
}
\description{
Transforming a VCF file into a dataframe.
}
